requirements.txt
# Core dependencies
ollama>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Document parsing
markdown>=3.5.0
PyYAML>=6.0.1

# Code execution
astunparse>=1.6.3

# CLI
click>=8.1.0
rich>=13.6.0

# Utilities
tqdm>=4.66.0
loguru>=0.7.0
